+++ NEWS.md
@@ -5,6 +5,9 @@
 * New format with only one block directory per archive uses less space,
   and breaks compatibility with 0.4.3.
 
+* New subcommands `conserve debug block list` and
+  `conserve debug block referenced`.
+
 ## Conserve 0.4.3 2018-10-13
 
 * `conserve versions` has a new `--sizes` option, to show disk usage by each